app/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.4
psycopg2==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart==0.0.6
boto3